feat: extract generic useCalendarList hook decoupled from Outlook
- Created apps/meteor/client/views/calendar/hooks/useCalendarList.ts
  with provider-agnostic useCalendarList(date) and useCalendarListForToday()
  hooks that call /v1/calendar-events.list endpoint directly
- Refactored useOutlookCalendarList to delegate to the generic hook
- Marked Outlook-specific wrappers as @deprecated
- Updated query keys from [outlook, calendar, list] to [calendar, list]
  so cache invalidation works across all calendar sources
- Aligns with GSoC goal of supporting multiple calendar integrations
